Massive sync against Debian work
This commit updates the APT repository to work with
http://apt.puppetlabs.com/prosvc/

In addition, the APT module from http://github.com/puppetlabs/puppet-apt
(which is a fork from a community module) is added as a submodule.

Finally, the module to manage repository sources on the nodes has been
replaced with equivalent logic and resources in the site module.