Make Vagrantfile more reliable.
This change fixes db* entries /etc/hosts, reduces the number of virtual
machines we're running at once, and turns on GUI boot mode.

These changes address some of the feedback that the vagrant project
doesn't work out of the box.

I'm going to need to use a settings system of sorts to make sure
package repositories are in order for the end user to just use this
vagrant project.