Add lucid mcollective server configuration
This is necessary to adjust the registration interval to something fast,
like every 2 seconds, for the live demonstration.

Normally, you would want something like 300 seconds, which is the
default in the upstream module.

NOTE: I have to override the entire configuration file, making it
default to keep the default client configuration file in sync with the
custom server configuration file.  This may be a good candidate for
hiera.